[BEAM-1871] Move ByteStringCoder to sdks/java/extensions/protobuf
